+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fix API link for `DrugStockReminderApi#drugStockReminder`
 - Check drug stock report status when app is opened
 - Schedule `DrugStock` API to run on 1st of every month at 8 AM
+- Migrate `ForgotPinCreateNewEffectHandler` to view effects
 
 ### Features